refactor(api): move sensor_metadata routes into separate file
